+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/Scroller` to show properly in RTL locales when `editable` is given
+
 ## [2.7.7] - 2023-08-22
 
 ### Added